reporting/requirements.txt
gspread==6.0.0
bottle==0.13.0
python-dotenv==1.0.1
